Added homeafrikalike.tk
Site with content that reads almost like gibberish, no valid backlinks. 

This is one of many referrer spam domains that uses "new-machina.html" in the URL.   If you Google "new-machina.html", a number of the results are replicated sites that use the same page theme/layout and many of the same images.